add ping() for win32
other changes include stub code for unsupported platforms, as
well as change in how modules are structured, in the sense, that
instead of declaring functions like _func_plat and then using a
macro to declare them as func, func is now declared as func itself.

if this doesn't make sense, see the diff. That should clear things up